Fixed addition of 100,000+ extra edges
Was adding the sibling edges in the ConnectGraph BFS *during* the BFS,
instead of after the BFS was complete. This resulted in a lot of extra
edges being added, which was causing poor runtime performance. I instead
buffered the edges to be added after the fact, which seems to have fixed
the issue!